Rename ktor-common-2.0 packages to match module name
