+++ instrumentation/ktor/ktor-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/ktor/v3_0/HttpClientInstrumentation.java
@@ -13,7 +13,7 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 import io.ktor.client.HttpClientConfig;
 import io.ktor.client.engine.HttpClientEngineConfig;
 import io.opentelemetry.api.GlobalOpenTelemetry;
-import io.opentelemetry.instrumentation.ktor.v2_0.common.internal.KtorBuilderUtil;
+import io.opentelemetry.instrumentation.ktor.common.v2_0.internal.KtorBuilderUtil;
 import io.opentelemetry.instrumentation.ktor.v3_0.KtorClientTelemetry;
 import io.opentelemetry.instrumentation.ktor.v3_0.KtorClientTelemetryBuilder;
 import io.opentelemetry.javaagent.bootstrap.internal.AgentCommonConfig;
