+++ instrumentation/ktor/ktor-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/ktor/v3_0/ServerInstrumentation.java
@@ -12,8 +12,8 @@ import io.ktor.server.application.Application;
 import io.ktor.server.application.ApplicationPluginKt;
 import io.ktor.server.engine.EmbeddedServer;
 import io.opentelemetry.api.GlobalOpenTelemetry;
-import io.opentelemetry.instrumentation.ktor.v2_0.common.AbstractKtorServerTelemetryBuilder;
-import io.opentelemetry.instrumentation.ktor.v2_0.common.internal.KtorBuilderUtil;
+import io.opentelemetry.instrumentation.ktor.common.v2_0.AbstractKtorServerTelemetryBuilder;
+import io.opentelemetry.instrumentation.ktor.common.v2_0.internal.KtorBuilderUtil;
 import io.opentelemetry.instrumentation.ktor.v3_0.KtorServerTelemetryBuilderKt;
 import io.opentelemetry.javaagent.bootstrap.internal.AgentCommonConfig;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
