+++ instrumentation/ktor/ktor-3.0/library/src/main/kotlin/io/opentelemetry/instrumentation/ktor/v3_0/KtorClientTelemetry.kt
@@ -12,8 +12,8 @@ import io.ktor.client.statement.*
 import io.ktor.util.*
 import io.opentelemetry.context.propagation.ContextPropagators
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter
-import io.opentelemetry.instrumentation.ktor.v2_0.common.AbstractKtorClientTelemetry
-import io.opentelemetry.instrumentation.ktor.v2_0.common.internal.KtorClientTelemetryUtil
+import io.opentelemetry.instrumentation.ktor.common.v2_0.AbstractKtorClientTelemetry
+import io.opentelemetry.instrumentation.ktor.common.v2_0.internal.KtorClientTelemetryUtil
 
 class KtorClientTelemetry internal constructor(
   instrumenter: Instrumenter<HttpRequestData, HttpResponse>,
